add socket and confirm of create, edit and remove

+++ my-project/src/views/Customer.vue
@@ -11,16 +11,16 @@
                         <div class="p-5 shadow-lg">
                             <div class="text-md font-raleway tracking-wide">
                                 <p>
-                                    Name: {{ item.attributes.Name }}
+                                    Name: {{ item.Name }}
                                 </p>
                                 <p>
-                                    Email: {{ item.attributes.Email }}
+                                    Email: {{ item.Email }}
                                 </p>
                                 <p>
-                                    Phone: {{ item.attributes.Phone }}
+                                    Phone: {{ item.Phone }}
                                 </p>
                                 <p>
-                                    Address: {{ item.attributes.Address }}
+                                    Address: {{ item.Address }}
                                 </p>
                                 <button class="flex-no-shrink p-2 mt-2 ml-2 border-2 rounded text-red border-red hover:bg-red"
                                     @click="edit(item.id)">Edit</button>
@@ -41,6 +41,8 @@
     import NavComp from '@/components/Nav.vue'
     import FooterComp from '@/components/Footer.vue'
     import PaginationComp from '@/components/Pagination'
+    import * as socketio from '../plugins/socketio';
+
     export default {
         name: 'CustomerPage',
         components: {
@@ -56,6 +58,16 @@
                 totalRecords: 0,
                 recordsPerPage: 3
             }
+        },
+        mounted() {
+            socketio.addEventListener({
+                type: 'customers',
+                callback: (eventData) => {
+                    this.customers = eventData.results
+                    this.totalPages = eventData.pagination.pageCount
+                    this.totalRecords = eventData.pagination.total
+                }
+            });
         },
         methods: {
             async loadListItem() {
@@ -74,20 +86,19 @@
                 this.loadListItem()
             },
             remove(cusId) {
-                this.axios.delete(`http://localhost:1337/api/customers/${cusId}`, {
-                    headers: {
-                        Authorization: `Bearer ${window.localStorage.getItem('jwt')}`,
-                    },
-                })
-                this.loadListItem()
+                if(confirm("Do you really want to delete?")) {
+                    this.axios.delete(`http://localhost:1337/api/customers/${cusId}`, {
+                        headers: {
+                            Authorization: `Bearer ${window.localStorage.getItem('jwt')}`,
+                        },
+                    })
+                    this.loadListItem()
+                }
             },
             edit(cusId) {
                 window.localStorage.setItem('cusId', cusId)
                 this.$router.push('/customer-item')
             }
-        },
-        created() {
-            this.loadListItem()
         }
     }
 </script>
